src/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
